pyproject.toml
[tool.poetry]
name = "nklf1"
version = "0.1.0"
description = ""
authors = ["PavelDusek <pavel.dusek@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.2"
flask-wtf = "^1.2.1"
wtforms = "^3.1.2"
numpy = "^1.26.4"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
